requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.0
Flask==3.0.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pluggy==1.3.0
pyflakes==3.1.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
ruff==0.1.11
six==1.16.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
PyYAML~=6.0.1
setuptools~=65.5.0